finance cal

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
-total_monthly_income = int(input("Enter your monthly income: "))
-total_monthly_expenses = int(input("Enter your total monthly expenses: "))
+total_monthly_income = float(input("Enter your monthly income: "))
+total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 
-monthly_saving = int(total_monthly_income - total_monthly_expenses)
+monthly_saving = float(total_monthly_income - total_monthly_expenses)
 
 savings = monthly_saving * 12 + (monthly_saving * 12 * 0.05)
